refactor the name of test file
