+++ aftership-sdk/src/test/java/com/aftership/sdk/endpoint/courier/ListAllCouriersTest.java
@@ -13,7 +13,7 @@ import com.aftership.sdk.rest.ResponseEntity;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
 
-public class TestListAllCouriers {
+public class ListAllCouriersTest {
     public static MockWebServer server;
 
     @BeforeAll
