+++ aftership-sdk/src/test/java/com/aftership/sdk/endpoint/notification/RemoveNotificationTest.java
@@ -1,4 +1,4 @@
-package com.aftership.sdk.endpoint.tracking;
+package com.aftership.sdk.endpoint.notification;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -7,15 +7,15 @@ import org.junit.jupiter.api.Test;
 import java.io.IOException;
 import com.aftership.sdk.AfterShip;
 import com.aftership.sdk.TestUtil;
-import com.aftership.sdk.model.tracking.MarkAsCompletedRequest;
-import com.aftership.sdk.model.tracking.MarkAsCompletedRequest.ReasonKind;
-import com.aftership.sdk.model.tracking.SingleTracking;
+import com.aftership.sdk.model.notification.NotificationWrapper;
 import com.aftership.sdk.model.tracking.SingleTrackingParam;
 import com.aftership.sdk.rest.DataEntity;
+import com.aftership.sdk.utils.JsonUtils;
+import com.aftership.sdk.utils.UrlUtils;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
 
-public class TestMarkAsCompleted {
+public class RemoveNotificationTest {
   public static MockWebServer server;
 
   @BeforeAll
@@ -23,7 +23,7 @@ public class TestMarkAsCompleted {
     server = new MockWebServer();
     server.enqueue(
         TestUtil.createMockResponse()
-            .setBody(TestUtil.getJson("endpoint/tracking/MarkAsCompletedResult.json")));
+            .setBody(TestUtil.getJson("endpoint/notification/RemoveNotificationResult.json")));
     server.start();
   }
 
@@ -33,29 +33,37 @@ public class TestMarkAsCompleted {
   }
 
   @Test
-  public void testMarkAsCompleted() throws IOException, InterruptedException {
+  public void testRemoveNotification() throws IOException, InterruptedException {
     AfterShip afterShip = TestUtil.createAfterShip(server);
 
     // request
     SingleTrackingParam param = new SingleTrackingParam();
     param.setId("100");
 
-    MarkAsCompletedRequest request = new MarkAsCompletedRequest(ReasonKind.LOST);
-    DataEntity<SingleTracking> entity =
-        afterShip.getTrackingEndpoint().markAsCompleted(param, request);
+    NotificationWrapper removedNotification =
+        JsonUtils.create()
+            .fromJson(
+                TestUtil.getJson("endpoint/notification/RemoveNotificationRequest.json"),
+                NotificationWrapper.class);
+    DataEntity<NotificationWrapper> entity =
+        afterShip.getNotificationEndpoint().removeNotification(param, removedNotification);
 
     // assert
     Assertions.assertFalse(entity.hasError(), "No errors in response.");
-    Assertions.assertNotNull(entity.getData().getTracking(), "Response data cannot be empty.");
+    Assertions.assertNotNull(entity.getData().getNotification(), "Response data cannot be empty.");
     Assertions.assertEquals(
-        "5b7658cec7c33c0e007de3c5", entity.getData().getTracking().getId(), "id  mismatch.");
+        2, entity.getData().getNotification().getEmails().length, "emails size mismatch.");
 
     RecordedRequest recordedRequest = server.takeRequest();
     Assertions.assertEquals("POST", recordedRequest.getMethod(), "Method mismatch.");
+    Assertions.assertEquals(
+        "/v4/notifications/100/remove",
+        UrlUtils.decode(recordedRequest.getPath()),
+        "path mismatch.");
 
     // output
     TestUtil.printResponse(afterShip, entity);
-    System.out.println("Path: " + recordedRequest.getPath());
+    System.out.println("Path: " + UrlUtils.decode(recordedRequest.getPath()));
     System.out.println("RequestBody: " + recordedRequest.getBody().readUtf8());
   }
 }
