+++ aftership-sdk/src/test/java/com/aftership/sdk/endpoint/tracking/CreateTrackingTest.java
@@ -1,22 +1,27 @@
 package com.aftership.sdk.endpoint.tracking;
 
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
 import java.io.IOException;
 import com.aftership.sdk.AfterShip;
 import com.aftership.sdk.TestUtil;
-import com.aftership.sdk.model.tracking.*;
+import com.aftership.sdk.model.tracking.CreateTrackingRequest;
+import com.aftership.sdk.model.tracking.SingleTracking;
 import com.aftership.sdk.rest.DataEntity;
 import com.aftership.sdk.utils.JsonUtils;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
 
-public class TestGetTracking {
+public class CreateTrackingTest {
     public static MockWebServer server;
 
     @BeforeAll
     static void setUp() throws IOException {
         server = new MockWebServer();
-        server.enqueue(TestUtil.createMockResponse().setBody(TestUtil.getJson("endpoint/tracking/GetTracking.json")));
+        server.enqueue(TestUtil.createMockResponse().setBody(TestUtil.getJson(
+                "endpoint/tracking/CreateTrackingResult.json")));
         server.start();
     }
 
@@ -26,29 +31,26 @@ public class TestGetTracking {
     }
 
     @Test
-    public void testGetTracking() throws IOException, InterruptedException {
+    public void testCreateTracking() throws IOException, InterruptedException {
         AfterShip afterShip = TestUtil.createAfterShip(server);
 
         //request
-        String getTrackingParams = TestUtil.getJson("endpoint/tracking/GetTrackingParams.json");
-        GetTrackingParams optionalParams = JsonUtils.create().fromJson(getTrackingParams, GetTrackingParams.class);
-        SingleTrackingParam param = new SingleTrackingParam();
-        param.setId("100");
-        param.setOptionalParams(new SingleTrackingOptionalParams());
-        param.getOptionalParams().setTrackingPostalCode("abc");
-        DataEntity<SingleTracking> entity = afterShip.getTrackingEndpoint().getTracking(param, optionalParams);
+        String requestBody = TestUtil.getJson("endpoint/tracking/CreateTrackingRequest.json");
+        CreateTrackingRequest request = JsonUtils.create().fromJson(requestBody, CreateTrackingRequest.class);
 
+        DataEntity<SingleTracking> entity = afterShip.getTrackingEndpoint().createTracking(request);
+
+        //assert
         Assertions.assertFalse(entity.hasError(), "No errors in response.");
         Assertions.assertNotNull(entity.getData().getTracking(), "Response data cannot be empty.");
-
         Assertions.assertEquals("fedex", entity.getData().getTracking().getSlug(), "Slug mismatch.");
-        Assertions.assertTrue(entity.getData().getTracking().getCheckpoints().size() > 0, "Checkpoints need to be " +
-                "greater than 0");
-        Assertions.assertEquals("InfoReceived", entity.getData().getTracking()
-                .getCheckpoints().get(0).getTag(), "tag mismatch.");
+        Assertions.assertEquals("2019-05-20", entity.getData().getTracking().getOrderPromisedDeliveryDate(),
+                "order_promised_delivery_date mismatch.");
+        Assertions.assertEquals(0, entity.getData().getTracking().getSmses().size(),
+                "Smses size mismatch.");
 
         RecordedRequest recordedRequest = server.takeRequest();
-        Assertions.assertEquals("GET", recordedRequest.getMethod(), "Method mismatch.");
+        Assertions.assertEquals("POST", recordedRequest.getMethod(), "Method mismatch.");
 
         //output
         TestUtil.printResponse(afterShip, entity);
