+++ aftership-sdk/src/test/java/com/aftership/sdk/endpoint/tracking/ReTrackTest.java
@@ -7,21 +7,20 @@ import org.junit.jupiter.api.Test;
 import java.io.IOException;
 import com.aftership.sdk.AfterShip;
 import com.aftership.sdk.TestUtil;
-import com.aftership.sdk.model.tracking.CreateTrackingRequest;
 import com.aftership.sdk.model.tracking.SingleTracking;
+import com.aftership.sdk.model.tracking.SingleTrackingParam;
 import com.aftership.sdk.rest.DataEntity;
-import com.aftership.sdk.utils.JsonUtils;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
 
-public class TestCreateTracking {
+public class ReTrackTest {
     public static MockWebServer server;
 
     @BeforeAll
     static void setUp() throws IOException {
         server = new MockWebServer();
         server.enqueue(TestUtil.createMockResponse().setBody(TestUtil.getJson(
-                "endpoint/tracking/CreateTrackingResult.json")));
+                "endpoint/tracking/ReTrackResult.json")));
         server.start();
     }
 
@@ -31,23 +30,18 @@ public class TestCreateTracking {
     }
 
     @Test
-    public void testCreateTracking() throws IOException, InterruptedException {
+    public void testReTrack() throws IOException, InterruptedException {
         AfterShip afterShip = TestUtil.createAfterShip(server);
 
-        //request
-        String requestBody = TestUtil.getJson("endpoint/tracking/CreateTrackingRequest.json");
-        CreateTrackingRequest request = JsonUtils.create().fromJson(requestBody, CreateTrackingRequest.class);
-
-        DataEntity<SingleTracking> entity = afterShip.getTrackingEndpoint().createTracking(request);
+        SingleTrackingParam param = new SingleTrackingParam();
+        param.setId("100");
+        DataEntity<SingleTracking> entity = afterShip.getTrackingEndpoint().reTrack(param);
 
         //assert
         Assertions.assertFalse(entity.hasError(), "No errors in response.");
         Assertions.assertNotNull(entity.getData().getTracking(), "Response data cannot be empty.");
-        Assertions.assertEquals("fedex", entity.getData().getTracking().getSlug(), "Slug mismatch.");
-        Assertions.assertEquals("2019-05-20", entity.getData().getTracking().getOrderPromisedDeliveryDate(),
-                "order_promised_delivery_date mismatch.");
-        Assertions.assertEquals(0, entity.getData().getTracking().getSmses().size(),
-                "Smses size mismatch.");
+        Assertions.assertTrue(entity.getData().getTracking().isActive(), "active mismatch.");
+        Assertions.assertEquals("russian-post", entity.getData().getTracking().getSlug(), "slug mismatch.");
 
         RecordedRequest recordedRequest = server.takeRequest();
         Assertions.assertEquals("POST", recordedRequest.getMethod(), "Method mismatch.");
@@ -57,5 +51,4 @@ public class TestCreateTracking {
         System.out.println("Path: " + recordedRequest.getPath());
         System.out.println("RequestBody: " + recordedRequest.getBody().readUtf8());
     }
-
 }
