+++ aftership-sdk/src/test/java/com/aftership/sdk/endpoint/tracking/UpdateTrackingTest.java
@@ -9,18 +9,20 @@ import com.aftership.sdk.AfterShip;
 import com.aftership.sdk.TestUtil;
 import com.aftership.sdk.model.tracking.SingleTracking;
 import com.aftership.sdk.model.tracking.SingleTrackingParam;
+import com.aftership.sdk.model.tracking.UpdateTrackingRequest;
 import com.aftership.sdk.rest.DataEntity;
+import com.aftership.sdk.utils.JsonUtils;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
 
-public class TestReTrack {
+public class UpdateTrackingTest {
     public static MockWebServer server;
 
     @BeforeAll
     static void setUp() throws IOException {
         server = new MockWebServer();
         server.enqueue(TestUtil.createMockResponse().setBody(TestUtil.getJson(
-                "endpoint/tracking/ReTrackResult.json")));
+                "endpoint/tracking/UpdateTrackingResult.json")));
         server.start();
     }
 
@@ -30,25 +32,31 @@ public class TestReTrack {
     }
 
     @Test
-    public void testReTrack() throws IOException, InterruptedException {
+    public void testTestUpdateTracking() throws IOException, InterruptedException {
         AfterShip afterShip = TestUtil.createAfterShip(server);
+
+        //request
+        String requestBody = TestUtil.getJson("endpoint/tracking/UpdateTrackingRequest.json");
+        UpdateTrackingRequest request = JsonUtils.create().fromJson(requestBody, UpdateTrackingRequest.class);
 
         SingleTrackingParam param = new SingleTrackingParam();
         param.setId("100");
-        DataEntity<SingleTracking> entity = afterShip.getTrackingEndpoint().reTrack(param);
+        DataEntity<SingleTracking> entity = afterShip.getTrackingEndpoint().updateTracking(param, request);
 
         //assert
         Assertions.assertFalse(entity.hasError(), "No errors in response.");
         Assertions.assertNotNull(entity.getData().getTracking(), "Response data cannot be empty.");
-        Assertions.assertTrue(entity.getData().getTracking().isActive(), "active mismatch.");
-        Assertions.assertEquals("russian-post", entity.getData().getTracking().getSlug(), "slug mismatch.");
+        Assertions.assertEquals("fedex", entity.getData().getTracking().getSlug(), "Slug mismatch.");
+        Assertions.assertEquals("InfoReceived", entity.getData().getTracking().getCheckpoints().get(0).getTag(),
+                "checkpoints::tag mismatch.");
 
         RecordedRequest recordedRequest = server.takeRequest();
-        Assertions.assertEquals("POST", recordedRequest.getMethod(), "Method mismatch.");
+        Assertions.assertEquals("PUT", recordedRequest.getMethod(), "Method mismatch.");
 
         //output
         TestUtil.printResponse(afterShip, entity);
         System.out.println("Path: " + recordedRequest.getPath());
         System.out.println("RequestBody: " + recordedRequest.getBody().readUtf8());
     }
+
 }
